// backend/src/modules/notifications/enums/NotificationType.ts
export enum NotificationType {
  OTP_LOGIN = 'OTP_LOGIN',
  OTP_WORK_START = 'OTP_WORK_START',
  WELCOME = 'WELCOME',
  BOOKING_CREATED = 'BOOKING_CREATED',
  BOOKING_CONFIRMED = 'BOOKING_CONFIRMED',
  WORK_STARTED = 'WORK_STARTED',
  WORK_COMPLETED = 'WORK_COMPLETED',
  PAYMENT_RECEIPT = 'PAYMENT_RECEIPT',
  BOOKING_CANCELLED = 'BOOKING_CANCELLED',
  REVIEW_REQUEST = 'REVIEW_REQUEST',
}

export enum NotificationChannel {
  EMAIL = 'EMAIL',
  SMS = 'SMS',
  PUSH = 'PUSH',
}